+++ src/index.html
@@ -50,7 +50,6 @@
   <article class="recipe-information-view hidden" id="recipeInfoView">
     <button class="favorite-button" id="favoritingButton">favorite this recipe</button>
     <button class="unfavorite-button hidden" id="unfavoritingButton">unfavorite this recipe</button>
-    <button class="go-shopping" id="goShopping">Go Shopping</button>
     <h2 id="recipeTitle">Recipe Title</h2> 
     <img class="current-recipe-image" id= "currentRecipeImage" src="./images/cookies.jpg" alt="food-image">
     <h3>Recipe ingredients</h3>
@@ -82,6 +81,7 @@
   <section class="recipes-to-cook hidden" id="recipesToCookView">
   </section>
   <section class="user-shopping-view hidden" id="userShoppingView">
+    <button class="lets-cook-button" id="letsCookButton">Let's Cook</button>
     <form>
       <label for="amt">Amount Needed:</label><br>
       <input type="text" id="amt" name="amt" required><br>
